Refactor Grid to become independent of CaseView
